Some improvements to vol.sh, it will work now even when pacmd can't connect to pulseaudio (no session daemon)

+++ vol.sh
@@ -7,9 +7,9 @@
 # Fernando Carmona Varo
 #
 
-# pulseaudio maximum volume value
+# pulseaudio maximum volume multiplier
 # Volume 65536 (0x10000) is normal max, greater values will amplify the audio signal (with clipping).
-max=0x20000
+multiplier=2
 
 percent=$1
 
@@ -21,47 +21,52 @@ fi
 
 if [ "$1" = "mute" ] # toggle mute
 then
-    if pacmd dump | grep "set-sink-mute .* yes" > /dev/null
-    then
-        pactl set-sink-mute 0 0 # unmute
-    else
-        pactl set-sink-mute 0 1 # mute
-    fi
+    pactl set-sink-mute @DEFAULT_SINK@ toggle
     exit
 fi
 
 
 current_vol=$(pacmd dump | awk '/set-sink-volume/ { print $3; exit }')
 
-set_vol=$((current_vol + (percent * max/100)))
-#set_vol=$((current_vol + (percent * 100)))
-
-
-if [ $((set_vol)) -lt $((0x0)) ]
+if [ "$current_vol" ]
 then
-    echo "Volume can't be set lower"
-    set_vol=$((0x0))
-else
-    if [ $((set_vol)) -gt $((max)) ]
+
+    max=$((multiplier * 0x10000))
+
+    set_vol=$((current_vol + (percent * max/100)))
+    #set_vol=$((current_vol + (percent * 100)))
+
+
+    if [ $((set_vol)) -lt $((0x0)) ]
     then
-        echo "Volume can't be set higher"; exit
-        set_vol=$((max))
+	echo "Volume can't be set lower"
+	set_vol=$((0x0))
+    else
+	if [ $((set_vol)) -gt $((max)) ]
+	then
+            echo "Volume can't be set higher"; exit
+            set_vol=$((max))
+	fi
     fi
+    display_vol=$((set_vol * 300/max))
+
+    pactl set-sink-volume @DEFAULT_SINK@ -- $set_vol
+
+
+    # notify with osd if available
+    hash osd_cat 2>$- && {
+	pkill osd_cat
+	osd_cat -O 3 -o 12 -c white -A center -d 1 -f "-*-*-*-*-*-*-*-*-*-*-*-*-*" \
+	    -b percentage -P $((set_vol * 100/max))
+    } &
+
+else
+    echo "Couldn't get current volume!"
+    display_vol="${percent}%"
+    pactl set-sink-volume @DEFAULT_SINK@ -- ${percent}%
 fi
-
-pactl set-sink-volume 0 $set_vol
-
-#display_vol=$((set_vol * 100/max))
-display_vol=$((set_vol * 300/max))
 
 # notify about the volume change in dwm
 xsetroot -name "volume: $display_vol" && sleep 1 && dwm.sh update &
-
-# notify with osd if available
-hash osd_cat 2>$- && {
-    pkill osd_cat
-    osd_cat -O 3 -o 12 -c white -A center -d 1 -f "-*-*-*-*-*-*-*-*-*-*-*-*-*" \
-	-b percentage -P $((set_vol * 100/max))
-} &
 
 echo $display_vol
